requirements.txt
anyio==3.6.1
astroid==2.11.7
bcrypt==3.2.2
black==22.6.0
cffi==1.15.1
click==8.1.3
dill==0.3.5.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
greenlet==1.1.2
h11==0.13.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.2
psycopg==3.0.16
psycopg2==2.9.3
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.9.2
pylint==2.14.5
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tomli==2.0.1
tomlkit==0.11.4
typing_extensions==4.3.0
uvicorn==0.18.2
wrapt==1.14.1
